split parser and datastore
The parser and datastore were both in parser before.
This commit split the datastore out to independent class
The split including following:
  parser -> parser, datastore
  featuresparser -> featuresparser and featuresdatastore
  fontparser -> fontparser and fontdatastore
  noteparser -> noteparser and notedatastore
  structuredtextparser -> structuredtextparser and structuredtextdatastore
