+++ src/parser/odbpp/fontparser.h
@@ -24,31 +24,7 @@
 #define __FONT_PARSER_H__
 
 #include "parser.h"
-#include "record.h"
-
-class FontDataStore: public DataStore {
-public:
-  void putXSize(const QStringList& param);
-  void putYSize(const QStringList& param);
-  void putOffset(const QStringList& param);
-  void charStart(const QStringList& param);
-  void charLineData(const QStringList& param);
-  void charEnd(void);
-
-  qreal offset(void);
-  qreal xsize(void);
-  qreal ysize(void);
-  CharRecord* charRecord(const char tchar);
-
-  virtual void dump(void);
-
-private:
-  qreal m_xsize, m_ysize;
-  qreal m_offset;
-  QMap<char, CharRecord*> m_records;
-
-  CharRecord* m_currentChar;
-};
+#include "fontdatastore.h"
 
 class FontParser: public Parser {
 public:
