+++ src/parser/odbpp/structuredtextparser.h
@@ -24,49 +24,7 @@
 #define __STRUCTURED_TEXT_PARSER_H__
 
 #include "parser.h"
-
-#include <string>
-
-using std::string;
-
-class StructuredTextDataStore: public DataStore {
-public:
-  class InvalidKeyException: public std::exception {
-  public:
-    InvalidKeyException(const char* msg): m_msg(msg) {}
-    virtual const char* what() const throw() { return m_msg; }
-
-  private:
-    const char* m_msg;
-  };
-
-  typedef enum { KEY_VALUE = 0, BLOCK } PutMode;
-
-  typedef map<string, string> ValueType;
-  typedef multimap<string, StructuredTextDataStore*> BlockType;
-  typedef BlockType::iterator BlockIter;
-  typedef pair<BlockIter, BlockIter> BlockIterPair;
-
-  StructuredTextDataStore();
-  void newElement(string name);
-  bool commitElement(void);
-
-  string get(string key);
-  BlockIterPair getBlocksByKey(string key);
-  const ValueType& getValueData(void);
-
-  void put(string key, string value);
-  virtual void dump(void);
-
-  static int dumpIndent;
-
-private:
-  PutMode m_mode;
-  ValueType m_valueData;
-  BlockType m_blockData;
-  StructuredTextDataStore* m_currentBlock;
-  string m_currentElementName;
-};
+#include "structuredtextdatastore.h"
 
 class StructuredTextParser : public Parser{
 public:
